--carry-on-with-disparities

+++ lib/linux.sh
@@ -80,6 +80,23 @@ lunux::unset_all_locales() {
 }
 
 linux::reset_locales() {
+  local carry_on=false
+
+  while [ "$#" -gt 0 ]; do
+    case "$1" in
+      -c|--carry-on-with-disparities)
+        carry_on=true
+        shift
+        ;;
+      -*)
+        softfail "Unknown argument: $1" || return $?
+        ;;
+      *)
+        break
+        ;;
+    esac
+  done
+
   local item; for item in "$@"; do
     if [[ "${item}" =~ ^[[:alpha:]_]+=(.*)$ ]]; then
       local locale_match; locale_match="$(<<<"${BASH_REMATCH[1]}" sed -E 's/[.]/[.]/g')" || softfail || return $?
@@ -98,24 +115,29 @@ linux::reset_locales() {
   #
   # <file>: line <line>: warning: setlocale: LC_<type>: cannot change locale (<locale name>): No such file or directory
   #
+  # In that case bash will not change it's own locale, as requested, but any child process started from that bash
+  # will use new locale as specified in environment.
+  #
   # The same command will produce no warnings on every consecutive run
   #
-  # I expect this peculiarity to be fixed before bash will get dead-code elimination, but if it's not going to happen then
-  # /usr/bin/true here is just for you guys from 2265, I hope that you are doing great there.
+  # I expect this peculiarity to be fixed before bash will get dead-code elimination, but if it's not going to happen
+  # then /usr/bin/true here is just for you guys from 2265, I hope that you are doing great there.
   #
   # printf '%(%c)T\n' could be used to test that case
   #
-  local temp_file; temp_file="$(mktemp)" || softfail || return $?
+  if [ "${carry_on}" = false ]; then
+    local temp_file; temp_file="$(mktemp)" || softfail || return $?
 
-  ( declare -gx "$@" && /usr/bin/true ) 2>"${temp_file}" || softfail || return $?
+    ( declare -gx "$@" && /usr/bin/true ) 2>"${temp_file}" || softfail || return $?
 
-  if [ -s "${temp_file}" ]; then
-    cat "${temp_file}" >&2 || softfail || return $?
-    rm "${temp_file}" || softfail || return $?
-    softfail "Unable to change locale, please try to run the same command once again in a new bash process"
-    return $?
-  else
-    rm "${temp_file}" || softfail || return $?
+    if [ -s "${temp_file}" ]; then
+      cat "${temp_file}" >&2 || softfail || return $?
+      rm "${temp_file}" || softfail || return $?
+      softfail "Unable to change locale, please try to run the same command once again in a new bash process"
+      return $?
+    else
+      rm "${temp_file}" || softfail || return $?
+    fi
   fi
 
   # -g global variable scope
